requirements.txt
requests
selenium
beautifulsoup4
tweepy
facebook-sdk
pandas
numpy
nltk
spacy
pymongo
psycopg2
sqlalchemy
scikit-learn
transformers
tensorflow
torch
sentence-transformers
keras
torchvision
scipy
hdbscan
kafka-python
pika
boto3
google-cloud-storage
firebase-admin
